Hide closed menus and relocate data modules
