+++ assets/data/currency.js
