+++ assets/data/equipment.d.ts
