+++ assets/data/equipment.js
